Rework POButtonStyle to conform to native ButtonStyle
