+++ Sources/ProcessOutCoreUI/Sources/DesignSystem/Button/POButtonStyle.swift
@@ -5,10 +5,10 @@
 //  Created by Andrii Vysotskyi on 22.11.2022.
 //
 
-import UIKit
+import SwiftUI
 
 /// Defines button style in all possible states.
-public struct POButtonStyle {
+public struct POButtonStyle: ButtonStyle {
 
     /// Style for normal state.
     public let normal: POButtonStateStyle
@@ -33,12 +33,54 @@ public struct POButtonStyle {
         self.disabled = disabled
         self.progressView = progressView
     }
+
+    // MARK: - ButtonStyle
+
+    public func makeBody(configuration: Configuration) -> some View {
+        let currentStyle = stateStyle(isPressed: configuration.isPressed)
+        ZStack {
+            POProgressView()
+                .progressViewStyle(progressView)
+                .opacity(isLoading ? 1 : 0)
+            configuration.label
+                .textStyle(currentStyle.title)
+                .multilineTextAlignment(.center)
+                .opacity(isLoading ? 0 : 1)
+        }
+        .padding(Constants.padding)
+        .frame(maxWidth: .infinity, minHeight: Constants.minHeight)
+        .background(Color(currentStyle.backgroundColor))
+        .border(style: currentStyle.border)
+        .shadow(style: currentStyle.shadow)
+        .allowsHitTesting(isEnabled && !isLoading)
+    }
+
+    // MARK: - Private Nested Types
+
+    private enum Constants {
+        static let minHeight: CGFloat = 44
+        static let padding = EdgeInsets(top: 4, leading: 8, bottom: 4, trailing: 8)
+    }
+
+    // MARK: - Private Properties
+
+    @Environment(\.isEnabled) private var isEnabled
+    @Environment(\.isButtonLoading) private var isLoading
+
+    // MARK: - Private Methods
+
+    private func stateStyle(isPressed: Bool) -> POButtonStateStyle {
+        if !isEnabled {
+            return disabled
+        }
+        return isPressed ? highlighted : normal
+    }
 }
 
 extension POButtonStyle {
 
     /// Default style for primary button.
-    public static let primary = POButtonStyle(
+    @_spi(PO) public static let primary = POButtonStyle(
         normal: .init(
             title: .init(color: UIColor(resource: .Text.on), typography: .Fixed.button),
             border: .clear,
@@ -61,7 +103,7 @@ extension POButtonStyle {
     )
 
     /// Default style for secondary button.
-    public static let secondary = POButtonStyle(
+    @_spi(PO) public static let secondary = POButtonStyle(
         normal: .init(
             title: .init(color: UIColor(resource: .Text.secondary), typography: .Fixed.button),
             border: .regular(color: UIColor(resource: .Border.default)),
